trigger watch task on livereload
With livereload set to true, watch task is triggered. Additional option
provided:
Name: files
Type: String|Array
This defines what file patterns this task will watch. Can be a string or
an array of files and/or minimatch patterns.iles

Type: String|Array

This defines what file patterns this task will watch. Can be a string or
an array of files and/or minimatch patterns.